+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## Unreleased
 
 - Add `fill` function and refactor `fillIn` to use it
-- Add event options bubbles and cancelable
+- Add change/input event options bubbles and cancelable
+- Add blur and improve focus/blur events
 
 ## 1.0.5 (10/9/19)